btrfs-progs: calc compressed size via search ioctl
Alternate way of calculating compressed size that does not need a new
ioctl and uses the SEARCH_TREE to traverse all inode extents.

The size of full extents is rounded up to 512 bytes, size of inline
extents is not. This gives reasonably accurate numbers to compute the
final ratio.

Notes:
* intended as a preview, not yet for inclusion
* the command is 'inspect-internal comprsize'
* the SEARCH_TREE ioctl is only available to root, thus not very useful
  to regular users and a separate ioctl (sent some time ago) may be
  still needed

Signed-off-by: David Sterba <dsterba@suse.cz>